move util functions to pointcloud_iof libraries
