+++ bbs3d/include/pointcloud_iof/pcl_eigen_coverter.hpp
@@ -1,0 +1,44 @@
+#pragma once
+#include <pcl/point_types.h>
+#include <pcl/point_cloud.h>
+#include <Eigen/Core>
+#include <vector>
+
+namespace pciof {
+template <typename T>
+void pcl_to_eigen(const pcl::PointCloud<pcl::PointXYZ>::Ptr& cloud_ptr, std::vector<T>& points) {
+  points.resize(cloud_ptr->size());
+  std::transform(cloud_ptr->begin(), cloud_ptr->end(), points.begin(), [](const pcl::PointXYZ& p) { return T(p.x, p.y, p.z); });
+}
+
+void eigen_to_pcl(const std::vector<Eigen::Vector3d>& points, pcl::PointCloud<pcl::PointXYZ>::Ptr& cloud_ptr) {
+  cloud_ptr->points.resize(points.size());
+  cloud_ptr->width = points.size();
+  cloud_ptr->height = 1;
+  cloud_ptr->is_dense = true;
+
+  std::transform(points.begin(), points.end(), cloud_ptr->points.begin(), [](const Eigen::Vector3d& v) {
+    pcl::PointXYZ point;
+    Eigen::Vector3f vf = v.cast<float>();
+    point.x = vf.x();
+    point.y = vf.y();
+    point.z = vf.z();
+    return point;
+  });
+}
+
+void eigen_to_pcl(const std::vector<Eigen::Vector3f>& points, pcl::PointCloud<pcl::PointXYZ>::Ptr& cloud_ptr) {
+  cloud_ptr->points.resize(points.size());
+  cloud_ptr->width = points.size();
+  cloud_ptr->height = 1;
+  cloud_ptr->is_dense = true;
+
+  std::transform(points.begin(), points.end(), cloud_ptr->points.begin(), [](const Eigen::Vector3f& v) {
+    pcl::PointXYZ point;
+    point.x = v.x();
+    point.y = v.y();
+    point.z = v.z();
+    return point;
+  });
+}
+}  // namespace pciof
