+++ ros2_test/config/ros2_test.yaml
@@ -1,6 +1,6 @@
 ################[Necessary]################
 ## Folder Paths
-target_clouds: "your_path/target"
+target_clouds: "/home/koki/Downloads/test_data/target"
 
 ## topic name
 lidar_topic_name: "/livox/points" ## msg type: sensor_msgs/PointCloud2
