+++ ros2_test/imgui.ini
@@ -1,10 +1,10 @@
 [Window][Debug##Default]
-Pos=706,552
+Pos=232,275
 Size=163,140
 Collapsed=0
 
 [Window][sub1]
-Pos=306,222
+Pos=53,-9
 Size=528,547
 Collapsed=0
