+++ ros2_test/rviz2/src/gpu_bbs3d_rviz2/gpu_ros2_test_rviz2.cpp
@@ -1,3 +1,5 @@
+#include <pointcloud_iof/pcl_eigen_coverter.hpp>
+#include <pointcloud_iof/gravity_align.hpp>
 #include <ros2_test_rviz2.hpp>
 #include <load_rviz2.hpp>
 #include <util.hpp>
@@ -58,7 +60,7 @@ ROS2Test::ROS2Test(const rclcpp::NodeOptions& node_options) : Node("gpu_ros2_tes
 
   // pcl to eigen
   std::vector<Eigen::Vector3f> tar_points;
-  pcl_to_eigen(tar_cloud_ptr, tar_points);
+  pciof::pcl_to_eigen(tar_cloud_ptr, tar_points);
 
   // broadcast viewer frame
   broadcast_viewer_frame(tar_points);
@@ -156,10 +158,12 @@ void ROS2Test::click_callback(const std_msgs::msg::Bool::SharedPtr msg) {
   }
 
   int imu_index = get_nearest_imu_index(imu_buffer, source_cloud_msg_->header.stamp);
-  gravity_align(src_cloud, src_cloud, imu_buffer[imu_index]);
+  const auto imu_msg = imu_buffer[imu_index];
+  const Eigen::Vector3d acc = {imu_msg.linear_acceleration.x, imu_msg.linear_acceleration.y, imu_msg.linear_acceleration.z};
+  pcl::transformPointCloud(*src_cloud, *src_cloud, pciof::calc_gravity_alignment_matrix(acc.cast<float>()));
 
   std::vector<Eigen::Vector3f> src_points;
-  pcl_to_eigen(src_cloud, src_points);
+  pciof::pcl_to_eigen(src_cloud, src_points);
   gpu_bbs3d.set_src_points(src_points);
 
   std::cout << "[Localize] start" << std::endl;
