+++ test/src/cpu_test.cpp
@@ -1,4 +1,6 @@
 #include <cpu_bbs3d/bbs3d.hpp>
+#include <pointcloud_iof/pcl_eigen_coverter.hpp>
+
 #include <test.hpp>
 #include <util.hpp>
 #include <load_yaml.hpp>
@@ -36,7 +38,7 @@ int BBS3DTest::run(std::string config) {
 
   // pcl to eigen
   std::vector<Eigen::Vector3d> tar_points;
-  pcl_to_eigen(tar_cloud_ptr, tar_points);
+  pciof::pcl_to_eigen(tar_cloud_ptr, tar_points);
 
   // Load source pcds
   std::cout << "[Setting] Loading source pcds..." << std::endl;
@@ -86,7 +88,7 @@ int BBS3DTest::run(std::string config) {
     std::cout << "-------------------------------" << std::endl;
     std::cout << "[Localize] pcd file name: " << src_cloud.first << std::endl;
     std::vector<Eigen::Vector3d> src_points;
-    pcl_to_eigen(src_cloud.second, src_points);
+    pciof::pcl_to_eigen(src_cloud.second, src_points);
     bbs3d_ptr->set_src_points(src_points);
     bbs3d_ptr->localize();
